Week-8-Intro-AgenticAI/requirements.txt
python-dotenv>=1.0.0
langchain-openai==1.1.0
langchain-core==1.1.0